refactor(model-class):reduce number of methods - create a super class for the model class - move some methods from model class to the super class [Finishes #37]
